# docs/openapi.yaml
openapi: 3.1.0
info:
  title: hushd API
  version: 0.7.0
  description: |
    hushd platform API: auth, verification, compliance (2257), billing, creator
    flows, media pipeline, messaging, payouts, moderation, and admin controls.

    ## Authentication

    Protected endpoints accept either:

    - `Authorization: Bearer <accessToken>` (API clients, tests), or
    - the `hushd_access` HttpOnly cookie set by `/auth/login`, `/auth/register`,
      `/auth/verify-email`, and `/auth/refresh` (browser frontends).

    Cookie-authenticated **mutating** requests must also send
    `X-Requested-With: XMLHttpRequest` (CSRF backstop). The refresh token
    travels in the `hushd_refresh` HttpOnly cookie scoped to `/v1/auth`;
    `/auth/refresh` and `/auth/logout` also accept it in the JSON body for
    non-browser clients.
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://localhost:3001/v1
    description: Local
...
paths:
...
  /verification/status:
    get:
      tags: [verification]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Get age/id verification status
      operationId: verificationStatus
      responses:
        "200":
          description: Verification status
  /verification/age/start:
    post:
      tags: [verification]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Start age verification session
      operationId: verificationAgeStart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartVerificationRequest"
      responses:
        "200":
          description: Verification session
  /verification/id/start:
    post:
      tags: [verification]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Start creator identity verification
      operationId: verificationIdStart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartVerificationRequest"
      responses:
        "200":
          description: Verification session
  /verification/admin/age/{verificationId}/approve:
    post:
      tags: [verification]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Manually approve an age verification (admin)
      operationId: verificationAdminAgeApprove
      parameters:
        - $ref: "#/components/parameters/verificationId"
      responses:
        "200":
          description: Approved
  /verification/admin/age/{verificationId}/reject:
    post:
      tags: [verification]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Manually reject an age verification (admin)
      operationId: verificationAdminAgeReject
      parameters:
        - $ref: "#/components/parameters/verificationId"
      responses:
        "200":
          description: Rejected
  /verification/admin/id/{verificationId}/approve:
    post:
      tags: [verification]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Manually approve an identity verification (admin)
      operationId: verificationAdminIdApprove
      parameters:
        - $ref: "#/components/parameters/verificationId"
      responses:
        "200":
          description: Approved
  /verification/admin/id/{verificationId}/reject:
    post:
      tags: [verification]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Manually reject an identity verification (admin)
      operationId: verificationAdminIdReject
      parameters:
        - $ref: "#/components/parameters/verificationId"
      responses:
        "200":
          description: Rejected
...
  /billing/subscriptions/checkout:
    post:
      tags: [billing]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Create processor checkout session
      operationId: billingCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionCheckoutRequest"
      responses:
        "200":
          description: Checkout session
  /billing/subscriptions:
    get:
      tags: [billing]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Own subscriptions
      operationId: billingSubscriptions
      responses:
        "200":
          description: Subscription rows
  /billing/subscriptions/cancel:
    post:
      tags: [billing]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Cancel subscription
      operationId: billingCancelSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelSubscriptionRequest"
      responses:
        "200":
          description: Canceled
  /billing/history:
    get:
      tags: [billing]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Payment history
      operationId: billingHistory
      responses:
        "200":
          description: Transaction rows
...
  /media/uploads/init:
    post:
      tags: [media]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Create staging upload target
      description: MIME type and byte size are required and validated before a presigned URL is issued.
      operationId: mediaUploadsInit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitUploadRequest"
      responses:
        "200":
          description: Presigned upload target
        "400":
          description: content_type_must_be_media | validation failed
  /media/uploads/complete:
    post:
      tags: [media]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Enqueue scan and ingest
      operationId: mediaUploadsComplete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteUploadRequest"
      responses:
        "200":
          description: Job accepted
  /media/{mediaId}:
    get:
      tags: [media]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Media asset status/details
      operationId: mediaGet
      parameters:
        - name: mediaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Media asset
        "404":
          description: Not found
  /messages/send:
    post:
      tags: [messages]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Send direct message (optional PPV)
      operationId: messagesSend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message created
  /messages/tips:
    post:
      tags: [messages]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Send a tip to a creator
      operationId: messagesTip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TipRequest"
      responses:
        "201":
          description: Tip recorded
  /messages/conversation/{creatorId}:
    get:
      tags: [messages]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Conversation with a creator
      operationId: messagesConversation
      parameters:
        - name: creatorId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Conversation with messages
  /messages/{messageId}/unlock:
    post:
      tags: [messages]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Unlock PPV message
      operationId: messagesUnlock
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Unlocked
...
  /moderation/reports:
    post:
      tags: [moderation]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: File a content/user report
      operationId: moderationReportCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportRequest"
      responses:
        "201":
          description: Report created
  /moderation/reports/me:
    get:
      tags: [moderation]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Own submitted reports
      operationId: moderationReportMine
      responses:
        "200":
          description: Reports
  /moderation/dmca:
    post:
      tags: [moderation]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: File a DMCA takedown notice (authenticated intake)
      operationId: moderationDmcaCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDmcaRequest"
      responses:
        "201":
          description: Notice recorded
  /moderation/queue:
    get:
      tags: [moderation]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Moderation queue (moderator/admin)
      operationId: moderationQueue
      responses:
        "200":
          description: Queue items
  /moderation/reports/{reportId}/triage:
    post:
      tags: [moderation]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Triage a report (moderator/admin)
      operationId: moderationReportTriage
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [TRIAGED, RESOLVED, DISMISSED] }
      responses:
        "200":
          description: Triaged
  /admin/moderation/queue:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Moderation queue for admins
      operationId: adminModerationQueue
      responses:
        "200":
          description: Queue items
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: hushd_access
      description: HttpOnly session cookie set by the auth endpoints.
  parameters:
    verificationId:
      name: verificationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    StartVerificationRequest:
      type: object
      required: [returnUrl]
      properties:
        returnUrl: { type: string, format: uri }
        locale: { type: string }
...
    CreateSubscriptionCheckoutRequest:
      type: object
      required: [creatorSlug, tierId, successReturnUrl, cancelReturnUrl]
      properties:
        creatorSlug: { type: string }
        tierId: { type: string }
        successReturnUrl: { type: string, format: uri }
        cancelReturnUrl: { type: string, format: uri }
    CancelSubscriptionRequest:
      type: object
      required: [subscriptionId]
      properties:
        subscriptionId: { type: string, format: uuid }
        reason: { type: string }
...
    InitUploadRequest:
      type: object
      required: [ownerCreatorId, mediaType, contentType, byteSize]
      properties:
        ownerCreatorId: { type: string, format: uuid }
        mediaType: { type: string, enum: [IMAGE, VIDEO, AUDIO, OTHER] }
        contentType: { type: string, pattern: "^(image|video|audio)/" }
        byteSize: { type: integer, minimum: 1, maximum: 524288000 }
    CompleteUploadRequest:
      type: object
      required: [mediaId, ownerCreatorId, stagingKey, mediaType]
      properties:
        mediaId: { type: string, format: uuid }
        ownerCreatorId: { type: string, format: uuid }
        stagingKey: { type: string }
        mediaType: { type: string, enum: [IMAGE, VIDEO, AUDIO, OTHER] }
        contentType: { type: string }
        sha256: { type: string }
    SendMessageRequest:
      type: object
      required: [creatorId]
      properties:
        creatorId: { type: string, format: uuid }
        fanUserId: { type: string, format: uuid, description: Required when the sender is the creator }
        body: { type: string }
        ppvPriceCents: { type: integer, minimum: 100, description: Creators only }
        currency: { type: string }
    TipRequest:
      type: object
      required: [creatorId, amountCents]
      properties:
        creatorId: { type: string, format: uuid }
        postId: { type: string, format: uuid }
        messageId: { type: string, format: uuid }
        amountCents: { type: integer, minimum: 100 }
        currency: { type: string }
...
    CreateReportRequest:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType: { type: string, enum: [POST, MESSAGE, USER, MEDIA] }
        targetId: { type: string }
        reason: { type: string }
    CreateDmcaRequest:
      type: object
      required: [claimantRef, targetRefs]
      properties:
        claimantRef: { type: string, minLength: 3, maxLength: 200 }
        targetRefs:
          type: array
          minItems: 1
          maxItems: 50
          items: { type: string }
        notesRef: { type: string, maxLength: 500 }
...